requirements.txt
django
channels[daphne]
psycopg2-binary
channels-redis